+++ src/App.js
@@ -1,13 +1,22 @@
-import Header from './components/Header';
-import Main from './components/Main';
+import {
+  BrowserRouter as Router,
+  Routes,
+  Route,
+} from 'react-router-dom';
+import Main from './pages/Main';
+import Contact from './pages/Contact';
 import Footer from './components/Footer';
 import './styles/App.css';
 
 function App() {
   return (
     <div className="App">
-      <Header />
-      <Main />
+      <Router>
+        <Routes>
+          <Route exact path="/" element={<Main />}></Route>
+          <Route path="/contact" element={<Contact />}></Route>
+        </Routes>
+      </Router>
       <Footer />
     </div>
   );
