+++ src/pages/Contact/index.js
@@ -1,0 +1,5 @@
+const Contact = () => {
+  return (<h1>Hello</h1>)
+};
+
+export default Contact;
